Add gRPC service to control the emulator
gRPC (https://grpc.io/) is a modern open source high performance RPC
framework that can run in any environment. It can efficiently connect
services in and across data centers with pluggable support for load
balancing, tracing, health checking and authentication.

We add a service definition that allows third party developers to
directly control the emulator by using the gRPC binding. The grpc
service can be launched by addig the -grpc <port> flag.

Currently it has the following limitations:

- Only available on X64 platforms
- No authentication
- No TLS

This will be used by:

- The webrtc module, which will need to control the emulator to forward
mouse/key and control events.
- Testing, the rpc bindings will enable easy testing scenarios.

A sample is added that demonstrates how to interact with the emulator
through an http bridge. See
android/android-grpc/docs/grpc-samples/GRPC.MD for details on how to use
it.

It makes some small changes to the compiler toolchain, and boring ssl to
make sure we can properly compile the depdendencies.

Change-Id: I1c79895d716c819ab719a1a42cd026bf3a1415b7