Enable gRPC for MSVC build.
The integration with android studio will require control of the emulator
through gRPC. We now enable gRPC for the MSVC build so we can control
the emulator on all supported platforms.  It remains disabled on MingW
due to lack of compiler support.

Contains some small fixes to make everything work on MSVC. Most notably
is the renaming of some variables due to protofub compiler not handling
"ERROR" label properly in windows.

Change-Id: I9dbab4dd0d157ca419f917383655453b3684b6ed
Bug: b/148079576