Always stream first frame and add frame counter
We now immediately send the first screen, vs. waiting for a change.

We now add a monotonically increasing frame number in a stream of
screenshots. The first screenshot will have a frame number of 0.
A single screenshot will always have a frame number of 0. The
frame is not necessarily contiguous, and can be used to detect how
many frames were dropped. An example frame sequence you might observe
is: [0, 3, 5, 7, 9, 11].

Bug: b/150503366
Test: sample/screenshot.py now streams immediately and displays frame
number.

Change-Id: Icabd356cfa2a753755514408cd163d503e60e97e