Add metrics around streamScreenshot
We now track the delivery time for frames. We track the 50, 95, 100%
percentile and the number of frames delivered.

We take this approach for the following reason:

The emulator only produces frames when the guest updates its screen.
Because of this we have variable frame rates. The best we can do is
track how long it takes to deliver a frame from the time we are notified
that a frame is available. This data will allow us:

- To calculate the number of frames that were dropped
- Get an idea of what frame rate we can deliver
- Get an idea of the time difference between the 50/95/100 percentile.

We include a timestamp in the protobuf Image so that android studio, the
WebRTC engine, and external UI can calculate the delay the introduced on
their side.

We extend the Percentile class to only serialize out buckets we care
about.

Change-Id: Id5afc943904d0f327c83e6f642b972355115a052