Derive coarse rotation from sensor state.
In our translation we ignored the x & y axis, and merely looked at the
z axis. A side effect of "moving" due to gpx playback is that the x-axis
gets set to -90 (imagine laying your phone flat on a table) and the
z-axis can change due to moving around.

We now use the same approach as used in the virtual sensors page to
derive the "coarse" device orientation, which should match how the OS
views the rotation.

Bug: b/174207107
Test: replay of gpx does not result in rotations.
Change-Id: I4e69b334817d68b2d9577a75d43e5d2f2e4ccd45