Fix audio down sampling
We incorrectly configured the SWVoice vs. the HWVoice. This caused
issues whenever there was a mismatch between the two.

We now make sure we do not touch the SWVoice, and modify the HWVoice
depending on what audio settings the user requested. Functions leaking
internal state have been removed, as they are no longer needed.

Note: It looks like there is a few ms of initial distortion, unclear why
that happens.

Test: Can record 8khz audio without distortion
Bug: 195497352
Change-Id: I0bb1cf63180f448c7804cb84ed04a8b8d46d16fd