Move GPS update mechanism into GPS driver
The flag `-no-passive-gps` toggles between one shot updates
vs. continious updates.

Now both the headless and UI version will use the same mechanism for
keeping the gps up to date.

This makes sure that the gRPC interface coordinates will not be
overriden by the location ui, which will stabilize the E2E tests where
we see flakiness due to the UI overriding the latest gps coordinates.

Note: Setting the gps coordinates through gRPC will not be reflected in
the UI.
Change-Id: If5294f95df43b98072eba8c991b0ac53ee10a4b0

Bug: 264296527
Change-Id: I78496596d4b0385cb859b3f87ed06a7d1f0d0727