Add support for android kernel events
This adds support for directly sending kernel events to the android
emulator.

This will allow developers to directly inject keycodes to trigger
behavior. For example you can now inject or eject a microphone. More
importantly it will enable us to transition more of the GUI to gRPC as
we will now be able to turn off the microphone etc..

Change-Id: I3177afdee957e22a9e9baf2bda100dcea415186a